move main.go to root to make main command name go-shamir on install
