+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/kinvolk/go-shamir/cli/cmd"
+import "github.com/kinvolk/go-shamir/cmd"
 
 func main() {
 	cmd.Execute()
